+++ CHANGELOG.md
@@ -11,6 +11,20 @@ NEXT-RELEASE Release notes (YYYY-MM-DD)
   Supported operators are `==`, `!=`, `>`, `<`, `>=`, `<=` and `contains(const std::string&)`.
 * Add `managed<std::map<std::string, T>>::contains_key` for conveniently checking if a managed map
   contains a given key. Use this method in the Type Safe Query API instead of `managed<std::map<std::string, T>>::find`.
+* Add `realm::networking` namespace which contains interfaces for providing your own custom network transport
+  implementations. The following interfaces are exposed:
+  * `websocket_interface`
+  * `websocket_observer`
+  * `sync_socket_provider`
+  * `sync_socket_provider::timer`
+  * `http_transport_client`
+* Add `default_http_transport` for built-in HTTP transport.
+* Add `default_socket_provider` a built-in class for providing the components necessary for transport via WebSocket.
+* A custom WebSocket & HTTP transport implementation can be used by passing 
+  the instance via `realm::app::App::configuration.http_transport_client` & `realm::app::App::configuration.sync_socket_provider`.
+* Proxy and custom header configuration for built-in transport must be supplied via the constructor on
+  `realm::networking::default_http_transport` & `realm::networking::default_socket_provider` using the
+  `realm::networking::default_transport_configuration` struct.
 
 ### Compatibility
 * Fileformat: Generates files with format v24. Reads and automatically upgrade from fileformat v10.